cast de int

+++ libft/ft_strncmp.c
@@ -6,7 +6,7 @@
 /*   By: mpressen <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2015/12/01 01:38:39 by mpressen          #+#    #+#             */
-/*   Updated: 2015/12/02 23:54:54 by mpressen         ###   ########.fr       */
+/*   Updated: 2015/12/05 06:03:52 by mpressen         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -19,8 +19,8 @@ int		ft_strncmp(const char *s1, const char *s2, size_t n)
 	count = 0;
 	while ((s1[count] || s2[count]) && count < n)
 	{
-		if (s1[count] != s2[count])
-			return (s1[count] - s2[count]);
+		if ((int)s1[count] != (int)s2[count])
+			return ((int)(s1[count] - s2[count]));
 		count++;
 	}
 	return (0);
